backend/requirements.txt
# Flask framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Database
mysql-connector-python>=8.2.0

# Authentication
Flask-Login>=0.6.3
PyJWT>=2.8.0
bcrypt>=4.1.0

# OAuth
Authlib>=1.3.0
requests>=2.31.0

# Async tasks
celery>=5.3.0
redis>=5.0.0

# Email
Flask-Mail>=0.9.1

# CORS
Flask-Cors>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0

# Date/time utilities
python-dateutil>=2.8.0
pytz>=2024.1

# External APIs
google-api-python-client>=2.110.0
google-auth>=2.25.0
facebook-sdk>=3.1.0
patreon>=0.5.0

# Validation
email-validator>=2.1.0

# Production server
gunicorn>=21.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=6.1.0
